requirements.txt
cloudinary==1.44.1
fastapi==0.124.4
passlib==1.7.4
pydantic==2.12.5
pydantic_settings==2.12.0
python-dotenv==1.2.1
python_jose==3.5.0
SQLAlchemy==2.0.45
bcrypt==4.3.0
